Stylesheet #1 Update
As of 4/23, a new portion is being added to index.html(Welcome Screen) which is currently a black box. This will house the first section of information regarding the website and its functionality. I plan to create multiple sections detailing the program's use and sample of what it can create. Along with animations for scrolling and coloring the sections an aesthetically pleasing color.